pyproject.toml
[project]
name = "llm-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dan Sclearov", email = "dansclearov@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.52.2",
    "colored>=2.3.0",
    "openai>=1.84.0",
    "platformdirs>=4.3.8",
    "requests>=2.32.3",
    "tenacity>=9.1.2",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "send2trash>=1.8.3",
    "prompt-toolkit>=3.0.51",
]


[project.scripts]
llm-cli = "llm_cli.main:main"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.0",
    "ruff>=0.13.2",
    "ty>=0.0.1a21",
]
